Refactor quiz components to use a shared base for reduced duplication
Create a `PracticeQuizBase` component to encapsulate shared quiz logic, including state management, UI rendering, and confetti animations. Refactor existing quiz components (`PracticeQuizUI`, `FactoryPracticeQuizUI`, `OceanPracticeQuizUI`) to utilize this base component, significantly reducing code duplication and improving maintainability.
