+++ client/src/components/game/OceanPracticeQuizUI.tsx
@@ -1,15 +1,5 @@
-import { useState, useEffect, useRef, useCallback } from "react";
 import { useGame } from "@/lib/stores/useGame";
-
-interface Question {
-  id: number;
-  code: string;
-  question: string;
-  options: string[];
-  correctIndex: number;
-  explanation: string;
-  hint: string;
-}
+import { PracticeQuizBase, type Question, type QuizTheme } from "./PracticeQuizBase";
 
 const QUESTIONS: Question[] = [
   {
@@ -138,321 +128,34 @@ while (roster.length > 0) {
   },
 ];
 
-interface ConfettiPiece {
-  id: number;
-  x: number;
-  y: number;
-  color: string;
-  size: number;
-  angle: number;
-  velocity: number;
-  spin: number;
-  delay: number;
-}
-
-const CONFETTI_COLORS = ["#4fc3f7", "#69f0ae", "#ffeb3b", "#ff9800", "#e040fb", "#ff6b6b", "#ffd93d", "#6bcb77"];
-
-function createConfetti(): ConfettiPiece[] {
-  const pieces: ConfettiPiece[] = [];
-  for (let i = 0; i < 40; i++) {
-    pieces.push({
-      id: i,
-      x: 50 + (Math.random() - 0.5) * 20,
-      y: 50,
-      color: CONFETTI_COLORS[Math.floor(Math.random() * CONFETTI_COLORS.length)],
-      size: 4 + Math.random() * 6,
-      angle: Math.random() * 360,
-      velocity: 80 + Math.random() * 120,
-      spin: (Math.random() - 0.5) * 720,
-      delay: Math.random() * 0.15,
-    });
-  }
-  return pieces;
-}
+const THEME: QuizTheme = {
+  title: "Ocean Practice Station",
+  accentColor: "#69f0ae",
+  hintColor: "#69f0ae",
+  bgColor: "rgba(0, 20, 50, 0.97)",
+  borderColor: "#69f0ae",
+  boxShadow: "0 0 40px rgba(105, 240, 174, 0.4)",
+  nextBtnTextColor: "#1b5e20",
+  confettiColors: ["#4fc3f7", "#69f0ae", "#ffeb3b", "#ff9800", "#e040fb", "#ff6b6b", "#ffd93d", "#6bcb77"],
+  confettiPrefix: "ocean-confetti",
+};
 
 export function OceanPracticeQuizUI() {
   const closeOceanPractice = useGame((s) => s.closeOceanPractice);
   const addOceanPracticeScore = useGame((s) => s.addOceanPracticeScore);
   const oceanPracticeScore = useGame((s) => s.oceanPracticeScore);
   const completeOceanPractice = useGame((s) => s.completeOceanPractice);
-
-  const [currentQ, setCurrentQ] = useState(0);
-  const [selectedAnswer, setSelectedAnswer] = useState<number | null>(null);
-  const [answeredCorrectly, setAnsweredCorrectly] = useState(false);
-  const [wrongAttempt, setWrongAttempt] = useState(false);
-  const [showHint, setShowHint] = useState(false);
-  const [confetti, setConfetti] = useState<ConfettiPiece[]>([]);
-  const [showConfetti, setShowConfetti] = useState(false);
-  const successSoundRef = useRef<HTMLAudioElement | null>(null);
-
-  useEffect(() => {
-    setCurrentQ(0);
-    setSelectedAnswer(null);
-    setAnsweredCorrectly(false);
-    setWrongAttempt(false);
-    setShowHint(false);
-    useGame.getState().resetOceanPracticeScore();
-    successSoundRef.current = new Audio("/sounds/success.mp3");
-    successSoundRef.current.volume = 0.5;
-  }, []);
-
-  const triggerCelebration = useCallback(() => {
-    if (successSoundRef.current) {
-      successSoundRef.current.currentTime = 0;
-      successSoundRef.current.play().catch(() => {});
-    }
-    setConfetti(createConfetti());
-    setShowConfetti(true);
-    setTimeout(() => setShowConfetti(false), 1200);
-  }, []);
-
-  const question = QUESTIONS[currentQ];
-  const isLastQuestion = currentQ >= QUESTIONS.length - 1;
-
-  const handleAnswer = (index: number) => {
-    if (answeredCorrectly) return;
-    setSelectedAnswer(index);
-    const correct = index === question.correctIndex;
-    if (correct) {
-      setAnsweredCorrectly(true);
-      addOceanPracticeScore(10);
-      triggerCelebration();
-    } else {
-      setWrongAttempt(true);
-      setTimeout(() => {
-        setSelectedAnswer(null);
-      }, 800);
-    }
-  };
-
-  const handleNext = () => {
-    if (isLastQuestion) {
-      completeOceanPractice();
-      return;
-    }
-    setCurrentQ((q) => q + 1);
-    setSelectedAnswer(null);
-    setAnsweredCorrectly(false);
-    setWrongAttempt(false);
-    setShowHint(false);
-  };
+  const resetOceanPracticeScore = useGame((s) => s.resetOceanPracticeScore);
 
   return (
-    <>
-    {showConfetti && (
-      <div style={{ position: "fixed", top: 0, left: 0, width: "100vw", height: "100vh", pointerEvents: "none", zIndex: 300, overflow: "hidden" }}>
-        {confetti.map((piece) => {
-          const rad = (piece.angle * Math.PI) / 180;
-          const endX = Math.cos(rad) * piece.velocity;
-          const endY = -Math.sin(rad) * piece.velocity + 200;
-          return (
-            <div
-              key={piece.id}
-              style={{
-                position: "absolute",
-                left: `${piece.x}%`,
-                top: `${piece.y}%`,
-                width: piece.size,
-                height: piece.size * 0.6,
-                background: piece.color,
-                borderRadius: piece.id % 3 === 0 ? "50%" : 2,
-                opacity: 1,
-                animation: `ocean-confetti-${piece.id} 1.1s ease-out ${piece.delay}s forwards`,
-              }}
-            >
-              <style>{`
-                @keyframes ocean-confetti-${piece.id} {
-                  0% { transform: translate(0, 0) rotate(0deg); opacity: 1; }
-                  80% { opacity: 1; }
-                  100% { transform: translate(${endX}px, ${endY}px) rotate(${piece.spin}deg); opacity: 0; }
-                }
-              `}</style>
-            </div>
-          );
-        })}
-      </div>
-    )}
-    <div
-      style={{
-        position: "absolute",
-        top: "50%",
-        left: "50%",
-        transform: "translate(-50%, -50%)",
-        background: "rgba(0, 20, 50, 0.97)",
-        borderRadius: 16,
-        padding: "24px 32px",
-        color: "white",
-        fontFamily: "'Inter', sans-serif",
-        zIndex: 200,
-        border: "3px solid #69f0ae",
-        boxShadow: "0 0 40px rgba(105, 240, 174, 0.4)",
-        width: 560,
-        maxHeight: "90vh",
-        overflowY: "auto",
-      }}
-    >
-      <div style={{ display: "flex", justifyContent: "space-between", alignItems: "center", marginBottom: 16 }}>
-        <div style={{ fontSize: 20, fontWeight: 800, color: "#69f0ae" }}>
-          Ocean Practice Station
-        </div>
-        <div style={{ display: "flex", gap: 16, alignItems: "center" }}>
-          <div style={{ fontSize: 14, color: "#ffeb3b", fontWeight: 700 }}>
-            Score: {oceanPracticeScore}
-          </div>
-          <div style={{ fontSize: 13, opacity: 0.6 }}>
-            {currentQ + 1} / {QUESTIONS.length}
-          </div>
-        </div>
-      </div>
-
-      <div style={{ fontSize: 14, fontWeight: 600, marginBottom: 8, color: "#b0bec5" }}>
-        {question.question}
-      </div>
-
-      <div
-        style={{
-          background: "rgba(0, 0, 0, 0.5)",
-          borderRadius: 8,
-          padding: "14px 18px",
-          fontFamily: "'Courier New', monospace",
-          fontSize: 13,
-          lineHeight: 1.7,
-          marginBottom: 16,
-          border: "1px solid rgba(105, 240, 174, 0.2)",
-          whiteSpace: "pre-wrap",
-          color: "#e0e0e0",
-        }}
-      >
-        {question.code}
-      </div>
-
-      <div style={{ display: "flex", flexDirection: "column", gap: 8, marginBottom: 16 }}>
-        {question.options.map((option, i) => {
-          let bg = "rgba(255,255,255,0.05)";
-          let border = "1px solid rgba(255,255,255,0.15)";
-          let textColor = "white";
-
-          if (answeredCorrectly && i === question.correctIndex) {
-            bg = "rgba(76, 175, 80, 0.3)";
-            border = "2px solid #66bb6a";
-            textColor = "#66bb6a";
-          } else if (selectedAnswer === i && !answeredCorrectly) {
-            bg = "rgba(244, 67, 54, 0.3)";
-            border = "2px solid #ef5350";
-            textColor = "#ef5350";
-          }
-
-          return (
-            <div
-              key={i}
-              onClick={() => handleAnswer(i)}
-              style={{
-                padding: "10px 16px",
-                background: bg,
-                border,
-                borderRadius: 8,
-                cursor: answeredCorrectly ? "default" : "pointer",
-                fontSize: 14,
-                color: textColor,
-                fontWeight: (answeredCorrectly && i === question.correctIndex) || selectedAnswer === i ? 600 : 400,
-                transition: "all 0.15s",
-              }}
-            >
-              {String.fromCharCode(65 + i)}) {option}
-            </div>
-          );
-        })}
-      </div>
-
-      {answeredCorrectly && (
-        <div
-          style={{
-            padding: "12px 16px",
-            background: "rgba(76, 175, 80, 0.15)",
-            border: "1px solid #66bb6a",
-            borderRadius: 8,
-            marginBottom: 16,
-          }}
-        >
-          <div style={{ fontSize: 15, fontWeight: 700, marginBottom: 4, color: "#66bb6a" }}>
-            Correct! +10 points
-          </div>
-          <div style={{ fontSize: 13, lineHeight: 1.6, opacity: 0.9 }}>
-            {question.explanation}
-          </div>
-        </div>
-      )}
-
-      {!answeredCorrectly && wrongAttempt && (
-        <div
-          style={{
-            padding: "12px 16px",
-            background: "rgba(244, 67, 54, 0.15)",
-            border: "1px solid #ef5350",
-            borderRadius: 8,
-            marginBottom: 16,
-          }}
-        >
-          <div style={{ fontSize: 15, fontWeight: 700, marginBottom: 4, color: "#ef5350" }}>
-            Not quite! Try again.
-          </div>
-          <div
-            onClick={() => setShowHint(!showHint)}
-            style={{
-              fontSize: 13,
-              color: "#69f0ae",
-              cursor: "pointer",
-              marginTop: 6,
-              display: "flex",
-              alignItems: "center",
-              gap: 6,
-            }}
-          >
-            <span style={{ fontSize: 10 }}>{showHint ? "\u25BC" : "\u25B6"}</span>
-            {showHint ? "Hide Hint" : "Show Hint"}
-          </div>
-          {showHint && (
-            <div style={{ fontSize: 13, lineHeight: 1.6, opacity: 0.85, marginTop: 8, paddingLeft: 16, borderLeft: "2px solid rgba(105, 240, 174, 0.4)" }}>
-              {question.hint}
-            </div>
-          )}
-        </div>
-      )}
-
-      <div style={{ display: "flex", justifyContent: "space-between", alignItems: "center" }}>
-        <div
-          onClick={closeOceanPractice}
-          style={{
-            padding: "8px 20px",
-            background: "rgba(255,255,255,0.1)",
-            border: "1px solid rgba(255,255,255,0.3)",
-            borderRadius: 8,
-            color: "white",
-            fontSize: 13,
-            cursor: "pointer",
-          }}
-        >
-          Exit
-        </div>
-        {answeredCorrectly && (
-          <div
-            onClick={handleNext}
-            style={{
-              padding: "10px 24px",
-              background: "#69f0ae",
-              border: "none",
-              borderRadius: 8,
-              color: "#1b5e20",
-              fontSize: 14,
-              fontWeight: 700,
-              cursor: "pointer",
-            }}
-          >
-            {isLastQuestion ? "Finish" : "Next Question"}
-          </div>
-        )}
-      </div>
-    </div>
-    </>
+    <PracticeQuizBase
+      questions={QUESTIONS}
+      theme={THEME}
+      score={oceanPracticeScore}
+      onClose={closeOceanPractice}
+      onAddScore={addOceanPracticeScore}
+      onResetScore={resetOceanPracticeScore}
+      onComplete={completeOceanPractice}
+    />
   );
 }
